requirements.txt
Flask
Flask-SocketIO==5.3.2
gunicorn==20.1.0
minimalmodbus==2.0.1
paho-mqtt==1.6.1
pyserial==3.5
